Fixes #3204, removed LOCK TABLES privilege check from installation. Modified LogDataPurger so log_actions are not purged if the lock tables privilege is not granted.
git-svn-id: http://dev.piwik.org/svn/trunk@6483 59fd770c-687e-43c8-a1e3-f5a4ff64c105

+++ plugins/PrivacyManager/LogDataPurger.php
@@ -73,8 +73,16 @@ class Piwik_PrivacyManager_LogDataPurger
 			}
 		}
 		
-		// delete unused actions from the log_action table
-		$this->purgeUnusedLogActions();
+		// delete unused actions from the log_action table (but only if we can lock tables)
+		if (Piwik::isLockPrivilegeGranted())
+		{
+			$this->purgeUnusedLogActions();
+		}
+		else
+		{
+			$logMessage = get_class($this).": LOCK TABLES privilege not granted; skipping unused actions purge";
+			Piwik::log($logMessage);
+		}
 		
 		// optimize table overhead after deletion
 		Piwik_OptimizeTables($logTables);
